Changes:
Implemented the service, and set it up with an alarm so that it runs in a given time interval.
https://app.asana.com/0/15290258811319/16145479436088
https://app.asana.com/0/15290258811319/16145479436090

The alarm starts at boot.
 https://app.asana.com/0/15290258811319/16146800712773

Set up a global class to hold global variables and flags.

Network service now updates the current list of applications to include new network traffic. TODO: fix update after device boot / resetting every day etc..
https://app.asana.com/0/15290258811319/16145479436125

The application state is now stored in sharedprefs between runs.
Changed the drawable field in Application to a string referencing the resource instead. This was to avoid having a circular dependency causing a runtimeExeption when serializing the object with Gson.
https://app.asana.com/0/15290258811319/16145479436121